# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ConnectInput {
  /** Phone number with country code, digits only (e.g. 905321234567) */
  phoneNumber?: string;
}

export type WhatsAppStatusState = typeof WhatsAppStatusState[keyof typeof WhatsAppStatusState];


export const WhatsAppStatusState = {
  disconnected: 'disconnected',
  connecting: 'connecting',
  qr_ready: 'qr_ready',
  pairing_code_ready: 'pairing_code_ready',
  connected: 'connected',
} as const;

export interface WhatsAppStatus {
  connected: boolean;
  state: WhatsAppStatusState;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  pairingCode?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  pushName?: string | null;
}

export type WhatsAppGroupType = typeof WhatsAppGroupType[keyof typeof WhatsAppGroupType];


export const WhatsAppGroupType = {
  group: 'group',
  channel: 'channel',
} as const;

export interface WhatsAppGroup {
  id: string;
  name: string;
  participantCount?: number;
  type: WhatsAppGroupType;
}

export interface SelectedGroups {
  groupIds: string[];
  groups?: WhatsAppGroup[];
}

export interface SelectedGroupsInput {
  groupIds: string[];
}

export interface FetchResult {
  status: string;
  message: string;
  count?: number;
}

export interface WhatsAppMessage {
  id: number;
  messageId?: string;
  groupId: string;
  groupName: string;
  content: string;
  sender: string;
  timestamp: string;
  fetchedAt?: string;
}

export interface MessagePage {
  messages: WhatsAppMessage[];
  total: number;
}

export interface ClearResult {
  deleted: number;
}

export type MessageStatsGroupsItem = {
  groupId: string;
  groupName: string;
  count: number;
};

export interface MessageStats {
  total: number;
  groups: MessageStatsGroupsItem[];
  listening: boolean;
  /** @nullable */
  lastFetchAt?: string | null;
  /** @nullable */
  nextFetchAt?: string | null;
}

export type GetMessagesParams = {
groupId?: string;
search?: string;
limit?: number;
offset?: number;
};