# lib/api-zod/src/generated/types/connectInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectInput {
  /** Phone number with country code, digits only (e.g. 905321234567) */
  phoneNumber?: string;
}