# lib/api-zod/src/generated/types/fetchResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */

export interface FetchResult {
  status: string;
  message: string;
  count?: number;
}