# lib/api-zod/src/generated/types/messagePage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppMessage } from './whatsAppMessage';

export interface MessagePage {
  messages: WhatsAppMessage[];
  total: number;
}