# lib/api-zod/src/generated/types/messageStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageStatsGroupsItem } from './messageStatsGroupsItem';

export interface MessageStats {
  total: number;
  groups: MessageStatsGroupsItem[];
  listening: boolean;
  /** @nullable */
  lastFetchAt?: string | null;
  /** @nullable */
  nextFetchAt?: string | null;
}