# lib/api-zod/src/generated/types/selectedGroups.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppGroup } from './whatsAppGroup';

export interface SelectedGroups {
  groupIds: string[];
  groups?: WhatsAppGroup[];
}