# lib/api-zod/src/generated/types/whatsAppGroup.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppGroupType } from './whatsAppGroupType';

export interface WhatsAppGroup {
  id: string;
  name: string;
  participantCount?: number;
  type: WhatsAppGroupType;
}