# lib/api-zod/src/generated/types/whatsAppMessage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppMessage {
  id: number;
  messageId?: string;
  groupId: string;
  groupName: string;
  content: string;
  sender: string;
  timestamp: string;
  fetchedAt?: string;
}