# lib/api-zod/src/generated/types/whatsAppStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp İlan Toplayıcı API
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsAppStatusState } from './whatsAppStatusState';

export interface WhatsAppStatus {
  connected: boolean;
  state: WhatsAppStatusState;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  pairingCode?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  pushName?: string | null;
}